backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Auth & security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.9

# Database (SQLite via SQLAlchemy)
sqlalchemy==2.0.29
aiosqlite==0.20.0

# HTTP / scraping
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.1
playwright==1.44.0        # optional: richer scraping; run `playwright install chromium`

# Document processing
python-docx==1.1.0
reportlab==4.1.0
PyPDF2==3.0.1
pdfplumber==0.11.0

# Excel
openpyxl==3.1.2

# AI
anthropic==0.25.3

# Email
aiosmtplib==3.0.1

# Utilities
python-dotenv==1.0.1
pydantic[email]==2.7.1
email-validator==2.1.1
pydantic-settings==2.2.1
aiofiles==23.2.1
